requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Django==4.0.6
idna==3.3
packaging==21.3
pgrep==2020.12.3
pyngrok==5.1.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
six
boto3==1.26.104
botocore==1.29.104
cloudinary==1.30.0
crispy-tailwind==0.5.0
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==1.2.2
django-mfa==3.2
django-otp==1.1.6
django-paypal==2.0
django-storages==1.13.2
django-two-factor-auth==1.15.1
pillow
python-dateutil==2.8.2
python-decouple==3.6
qrcode==7.4.2
phonenumbers
s3transfer==0.6.0
simplejson==3.18.3
coinbase-commerce
paystack
django-mysql
